Move reference data to .csv files

+++ src/Database/Membership/Tables/UserRole.sql
@@ -1,5 +1,5 @@
 ﻿CREATE TABLE [Membership].[UserRole] (
-    [RoleID] INT          NOT NULL,
+    [RoleID] TINYINT      NOT NULL,
     [Name]   [dbo].[Name] NOT NULL,
     CONSTRAINT [PK_UserRole_RoleID] PRIMARY KEY CLUSTERED ([RoleID] ASC)
 );
