+++ src/Database/Membership/Tables/UserUserRole.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TABLE [Membership].[UserUserRole] (
-    [UserID] INT NOT NULL,
-    [RoleID] INT NOT NULL,
+    [UserID] INT     NOT NULL,
+    [RoleID] TINYINT NOT NULL,
     CONSTRAINT [PK_UserUserRole_UserID_RoleID] PRIMARY KEY CLUSTERED ([UserID] ASC, [RoleID] ASC),
     CONSTRAINT [FK_UserUserRole_User] FOREIGN KEY ([UserID]) REFERENCES [Membership].[User] ([UserID]) ON DELETE CASCADE ON UPDATE CASCADE,
     CONSTRAINT [FK_UserUserRole_UserRole] FOREIGN KEY ([RoleID]) REFERENCES [Membership].[UserRole] ([RoleID]) ON DELETE CASCADE ON UPDATE CASCADE
